\name{x2}
\alias{x2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to return the product of the argument by it self.
}
\description{
x
}
\usage{
x2(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x
}
}
\details{
x
}
\value{
x
}
\references{
x
}
\author{
Cristiano
}
\note{
x
}
